moved WidthModifier classes to dedicated module
