+++ autofit/__init__.py
@@ -19,14 +19,14 @@ from .mapper.prior.assertion import GreaterThanLessThanAssertion
 from .mapper.prior.assertion import GreaterThanLessThanEqualAssertion
 from .mapper.prior.deferred import DeferredArgument
 from .mapper.prior.deferred import DeferredInstance
-from .mapper.prior.prior import AbsoluteWidthModifier
+from .mapper.prior.width_modifier import AbsoluteWidthModifier
+from .mapper.prior.width_modifier import RelativeWidthModifier
+from .mapper.prior.width_modifier import WidthModifier
 from .mapper.prior.prior import GaussianPrior
 from .mapper.prior.prior import LogUniformPrior
 from .mapper.prior.prior import Prior
-from .mapper.prior.prior import RelativeWidthModifier
 from .mapper.prior.prior import TuplePrior
 from .mapper.prior.prior import UniformPrior
-from .mapper.prior.prior import WidthModifier
 from .mapper.prior_model.abstract import AbstractPriorModel
 from .mapper.prior_model.annotation import AnnotationPriorModel
 from .mapper.prior_model.attribute_pair import AttributeNameValue
